[feature: #163845211] Restructure files and folders move all images to img directory move all component styles to respective component module refactore the import and export routes remove old snapshots and recreate new ones [Finishes: 163845211]
