+++ src/app/footer/footer.scss
