+++ src/app/home/home.scss
@@ -10,7 +10,7 @@ h4 {
 #landingPage {
   height: 85%;
   width: auto;
-  background-image: url('../img/landingPageImage.jpeg');
+  background-image: url('../../img/landingPageImage.jpeg');
   background-size: cover;
   position: relative;
   background-position: center center;
